# specs/002-kiosk-screen/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Kiosk Screen API
  version: 0.1.0
  description: MVP API contract for kiosk display, content, ads, readiness, roles, and approved embedded domains.
servers:
  - url: /api
security:
  - sessionCookie: []
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: kiosk_session
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 1
    User:
      type: object
      required: [id, email, displayName, roles]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        displayName:
          type: string
        roles:
          type: array
          items:
            type: string
            enum: [display_viewer, event_operator, content_manager, advertising_manager, administrator]
    KioskConfiguration:
      type: object
      required: [id, name, topRegionRatio, bottomRegionRatio, defaultTopDurationSeconds, defaultAdDurationSeconds, configuredEventDurationMinutes, isEnabled]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        topRegionRatio:
          type: integer
          const: 4
        bottomRegionRatio:
          type: integer
          const: 1
        defaultTopDurationSeconds:
          type: integer
          minimum: 1
        defaultAdDurationSeconds:
          type: integer
          minimum: 1
        configuredEventDurationMinutes:
          type: integer
          minimum: 1
        isEnabled:
          type: boolean
    KioskConfigurationRequest:
      type: object
      required: [name, defaultTopDurationSeconds, defaultAdDurationSeconds, configuredEventDurationMinutes, isEnabled]
      properties:
        name:
          type: string
        defaultTopDurationSeconds:
          type: integer
          minimum: 1
        defaultAdDurationSeconds:
          type: integer
          minimum: 1
        configuredEventDurationMinutes:
          type: integer
          minimum: 1
        isEnabled:
          type: boolean
...
    Client:
      type: object
      required: [id, name, isActive]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        isActive:
          type: boolean
    ClientRequest:
      type: object
      required: [name, isActive]
      properties:
        name:
          type: string
        isActive:
          type: boolean
...
    ApprovedEmbeddedDomain:
      type: object
      required: [id, domain, isActive]
      properties:
        id:
          type: string
          format: uuid
        domain:
          type: string
        isActive:
          type: boolean
    ApprovedEmbeddedDomainRequest:
      type: object
      required: [domain, isActive]
      properties:
        domain:
          type: string
        isActive:
          type: boolean
...
paths:
  /auth/login:
    post:
      tags: [Authentication]
      summary: Start an authenticated session.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Invalid credentials.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /auth/me:
    get:
      tags: [Authentication]
      summary: Read current authenticated user.
      responses:
        '200':
          description: Current user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /display/configuration:
    get:
      tags: [Display]
      summary: Read kiosk display configuration.
      responses:
        '200':
          description: Kiosk configuration.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KioskConfiguration'
    put:
      tags: [Display]
      summary: Update kiosk display configuration.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KioskConfigurationRequest'
      responses:
        '200':
          description: Updated kiosk configuration.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KioskConfiguration'
        '403':
          description: Administrator role required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /clients:
    get:
      tags: [Clients]
      summary: List clients.
      responses:
        '200':
          description: Clients.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Client'
    post:
      tags: [Clients]
      summary: Create client.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientRequest'
      responses:
        '201':
          description: Created client.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
  /clients/{clientId}:
    parameters:
      - name: clientId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    put:
      tags: [Clients]
      summary: Update client.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientRequest'
      responses:
        '200':
          description: Updated client.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
    delete:
      tags: [Clients]
      summary: Remove client when no active ads depend on it.
      responses:
        '204':
          description: Removed.
        '409':
          description: Active ads depend on this client.
...
  /approved-domains:
    get:
      tags: [Approved Domains]
      summary: List approved embedded domains.
      responses:
        '200':
          description: Approved embedded domains.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ApprovedEmbeddedDomain'
    post:
      tags: [Approved Domains]
      summary: Create approved embedded domain.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApprovedEmbeddedDomainRequest'
      responses:
        '201':
          description: Created approved domain.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovedEmbeddedDomain'
  /approved-domains/{domainId}:
    parameters:
      - name: domainId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    put:
      tags: [Approved Domains]
      summary: Update approved embedded domain.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApprovedEmbeddedDomainRequest'
      responses:
        '200':
          description: Updated approved domain.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovedEmbeddedDomain'
    delete:
      tags: [Approved Domains]
      summary: Remove approved embedded domain when no active iframe content depends on it.
      responses:
        '204':
          description: Removed.
        '409':
          description: Active iframe content depends on this domain.
...